@@ -17,7 +17,7 @@
         "symfony/twig-bundle": "^2.4|^3.0|^4.0",
         "symfony/http-foundation": "^2.4|^3.0|^4.0",
         "symfony/options-resolver": "^2.6|^3.0|^4.0",
-        "symfony/form" : "^2.4|^3.0|^4.0",
+        "symfony/form" : "^2.7.39|^3.0|^4.0",
         "symfony/validator": "^2.4|^3.0|^4.0",
         "doctrine/doctrine-bundle": "^1.2",
         "fsi/doctrine-extensions": "^2.0@dev",
@@ -25,13 +25,13 @@
         "knplabs/knp-gaufrette-bundle": "~0.1"
     },
     "require-dev": {
-        "bossa/phpspec2-expect": "^2.3",
+        "bossa/phpspec2-expect": "^2.3|^3.0",
         "doctrine/orm": "^2.5",
         "fsi/datagrid-bundle": "^1.1|^2.0@dev",
-        "phpunit/phpunit": "^4.8|^5.7",
-        "phpspec/phpspec": "^3.0",
+        "phpunit/phpunit": "^6.5",
+        "phpspec/phpspec": "^3.0|^4.3",
         "symfony/asset": "^2.7|^3.0|^4.0",
-        "symfony/phpunit-bridge": "^3.3|^4.0",
+        "symfony/phpunit-bridge": "^4.0",
         "symfony/security": "^2.4|^3.0|^4.0",
         "symfony/twig-bridge": "^2.4|^3.0|^4.0"
     },
