@@ -7,11 +7,18 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace FSi\Bundle\DoctrineExtensionsBundle\Tests\Form\Type\FSi;
 
+use FSi\Bundle\DoctrineExtensionsBundle\Form\Type\FSi\FileType;
+use FSi\Bundle\DoctrineExtensionsBundle\Listener\Uploadable\Filesystem;
 use FSi\Bundle\DoctrineExtensionsBundle\Resolver\FSiFilePathResolver;
 use FSi\Bundle\DoctrineExtensionsBundle\Tests\Fixtures\Entity\Article;
 use FSi\Bundle\DoctrineExtensionsBundle\Tests\Fixtures\Form\Extension\FSiFileExtension;
+use FSi\DoctrineExtensions\Uploadable\File;
+use Gaufrette\Adapter\Local;
+use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
@@ -34,12 +41,8 @@ class FileTypeTest extends FormTypeTest
         $article = new Article();
         $article->setFile($file);
 
-        $form = $this->factory->create(
-            $this->isSymfony3() ? 'Symfony\Component\Form\Extension\Core\Type\FormType' : 'form',
-            $article
-        );
-
-        $form->add('file', $this->isSymfony3() ? 'FSi\Bundle\DoctrineExtensionsBundle\Form\Type\FSi\FileType' : 'fsi_file');
+        $form = $this->factory->create($this->isSymfony3() ? FormType::class : 'form', $article);
+        $form->add('file', $this->isSymfony3() ? FileType::class : 'fsi_file');
 
         $html = $this->twig->render('form_with_fsi_file.html.twig', [
             'form' => $form->createView()
@@ -55,11 +58,11 @@ class FileTypeTest extends FormTypeTest
         $article->setFile($file);
 
         $form = $this->factory->create(
-            $this->isSymfony3() ? 'Symfony\Component\Form\Extension\Core\Type\FormType' : 'form',
+            $this->isSymfony3() ? FormType::class : 'form',
             $article
         );
 
-        $form->add('file', $this->isSymfony3() ? 'FSi\Bundle\DoctrineExtensionsBundle\Form\Type\FSi\FileType' : 'fsi_file', [
+        $form->add('file', $this->isSymfony3() ? FileType::class : 'fsi_file', [
             'file_url' => function (UrlGeneratorInterface $urlGenerator, FormInterface $form) {
                 return 'constant_file_url';
             },
@@ -71,35 +74,37 @@ class FileTypeTest extends FormTypeTest
         $this->assertSame($html, $this->getExpectedHtml('form_with_fsi_file_and_constant_url.html'));
     }
 
-    private function getFileMock(): \PHPUnit_Framework_MockObject_MockObject
+    private function getFileMock()
     {
-        $file = $this->getMockBuilder('FSi\DoctrineExtensions\Uploadable\File')
+        $file = $this->getMockBuilder(File::class)
             ->disableOriginalConstructor()
-            ->getMock();
+            ->getMock()
+        ;
 
         $file->expects($this->any())
             ->method('getFilesystem')
             ->will(
                 $this->returnCallback(
                     function () {
-                        $fileSystem = $this->getMockBuilder(
-                            'FSi\Bundle\DoctrineExtensionsBundle\Listener\Uploadable\Filesystem'
-                        )
+                        $fileSystem = $this->getMockBuilder(Filesystem::class)
                             ->disableOriginalConstructor()
-                            ->getMock();
+                            ->getMock()
+                        ;
 
                         $fileSystem->expects($this->any())
                             ->method('getBaseUrl')
-                            ->will($this->returnValue('/uploaded/'));
+                            ->will($this->returnValue('/uploaded/'))
+                        ;
 
                         $fileSystem->expects($this->any())
                             ->method('getAdapter')
                             ->will(
                                 $this->returnCallback(
                                     function () {
-                                        return $this->getMockBuilder('Gaufrette\Adapter\Local')
+                                        return $this->getMockBuilder(Local::class)
                                             ->disableOriginalConstructor()
-                                            ->getMock();
+                                            ->getMock()
+                                        ;
                                     }
                                 )
                             );
@@ -111,11 +116,13 @@ class FileTypeTest extends FormTypeTest
 
         $file->expects($this->any())
             ->method('getKey')
-            ->will($this->returnValue('Article/file/1/image name.jpg'));
+            ->will($this->returnValue('Article/file/1/image name.jpg'))
+        ;
 
         $file->expects($this->any())
             ->method('getName')
-            ->will($this->returnValue('Article/file/1/image name.jpg'));
+            ->will($this->returnValue('Article/file/1/image name.jpg'))
+        ;
 
         return $file;
     }
