@@ -7,10 +7,13 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace FSi\Bundle\DoctrineExtensionsBundle\Tests\Form\Type\FSi;
 
 use FSi\Bundle\DoctrineExtensionsBundle\Resolver\FSiFilePathResolver;
 use FSi\Bundle\DoctrineExtensionsBundle\Twig\FilesExtension;
+use RuntimeException;
 use Symfony\Bridge\Twig\Command\DebugCommand;
 use Symfony\Bridge\Twig\Extension\AssetExtension;
 use Symfony\Bridge\Twig\Extension\FormExtension;
@@ -19,20 +22,20 @@ use Symfony\Bridge\Twig\Form\TwigRenderer;
 use Symfony\Bridge\Twig\Form\TwigRendererEngine;
 use Symfony\Bridge\Twig\Tests\Extension\Fixtures\StubFilesystemLoader;
 use Symfony\Bridge\Twig\Tests\Extension\Fixtures\StubTranslator;
-use Symfony\Bundle\TwigBundle\Extension\AssetsExtension;
 use Symfony\Component\Asset\Packages;
 use Symfony\Component\Asset\UrlPackage;
 use Symfony\Component\Asset\VersionStrategy\EmptyVersionStrategy;
+use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormRenderer;
-// prior to 2.7 asset component was part of FrameworkBundle
-use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
-use Symfony\Component\Templating\Asset\UrlPackage as LegacyUrlPackage;
 use Symfony\Component\Form\Test\FormIntegrationTestCase;
+use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
+use Twig_Environment;
+use Twig_RuntimeLoaderInterface;
 
 abstract class FormTypeTest extends FormIntegrationTestCase
 {
     /**
-     * @var \Twig_Environment
+     * @var Twig_Environment
      */
     protected $twig;
 
@@ -50,7 +53,7 @@ abstract class FormTypeTest extends FormIntegrationTestCase
 
         $loader = new StubFilesystemLoader($paths);
 
-        $twig = new \Twig_Environment($loader, ['strict_variables' => true]);
+        $twig = new Twig_Environment($loader, ['strict_variables' => true]);
         $twig->addGlobal('global', '');
         $twig->addExtension(new TranslationExtension(new StubTranslator()));
         $twig->addExtension(new FilesExtension(new FSiFilePathResolver()));
@@ -67,51 +70,35 @@ abstract class FormTypeTest extends FormIntegrationTestCase
             );
         }
 
-        if (class_exists('Symfony\Bridge\Twig\Extension\AssetExtension')) {
-            $runtimeLoader = $this->getMockBuilder('Twig_RuntimeLoaderInterface')->getMock();
-            $runtimeLoader->expects($this->any())->method('load')->will($this->returnValueMap([
-                [TwigRenderer::class, $renderer],
-                [FormRenderer::class, $renderer],
-            ]));
-            $twig->addRuntimeLoader($runtimeLoader);
-
-            $twig->addExtension(new FormExtension());
-            $twig->addExtension(new AssetExtension(
-                new Packages(
-                    new UrlPackage(['http://local.dev/'], new EmptyVersionStrategy())
-                )
-            ));
-        } else {
-            $twig->addExtension(new FormExtension($renderer));
-            $container = $this->getMockBuilder('Symfony\Component\DependencyInjection\ContainerInterface')->getMock();
-            $container->expects($this->any())
-                ->method('get')
-                ->with('templating.helper.assets')
-                ->will($this->returnValue(new LegacyUrlPackage()));
-
-            $request = $this->getMockBuilder('Symfony\Component\Routing\RequestContext')->getMock();
-            $twig->addExtension(new AssetsExtension($container, $request));
-        }
+        $runtimeLoader = $this->getMockBuilder(Twig_RuntimeLoaderInterface::class)->getMock();
+        $runtimeLoader->expects($this->any())->method('load')->will($this->returnValueMap([
+            [TwigRenderer::class, $renderer],
+            [FormRenderer::class, $renderer],
+        ]));
+        $twig->addRuntimeLoader($runtimeLoader);
+        $twig->addExtension(new FormExtension($renderer));
+        $twig->addExtension(new AssetExtension(
+            new Packages(new UrlPackage(['http://local.dev/'], new EmptyVersionStrategy()))
+        ));
 
         $this->twig = $twig;
     }
 
     /**
-     * @param string $filename
-     * @return string
+     * @throws RuntimeException
      */
-    protected function getExpectedHtml($filename)
+    protected function getExpectedHtml(string $filename): string
     {
         $path = __DIR__ . '/../../../Resources/views/' . $filename;
         if (!file_exists($path)) {
-            throw new \RuntimeException(sprintf('Invalid expected html file path "%s"', $path));
+            throw new RuntimeException(sprintf('Invalid expected html file path "%s"', $path));
         }
 
         return trim(file_get_contents($path));
     }
 
-    protected function isSymfony3()
+    protected function isSymfony3(): bool
     {
-        return method_exists('Symfony\Component\Form\AbstractType', 'getBlockPrefix');
+        return method_exists(AbstractType::class, 'getBlockPrefix');
     }
 }
