@@ -1,8 +1,16 @@
 <?php
 
+/**
+ * (c) FSi sp. z o.o. <info@fsi.pl>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
 namespace FSi\Bundle\DoctrineExtensionsBundle\Validator\Constraints\Symfony3;
 
 final class ImageValidator extends FileValidator
 {
-
 }
