@@ -1,9 +1,19 @@
 <?php
 
+/**
+ * (c) FSi sp. z o.o. <info@fsi.pl>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
 namespace spec\FSi\Bundle\DoctrineExtensionsBundle\EventListener;
 
 use FSi\DoctrineExtensions\Translatable\TranslatableListener;
 use PhpSpec\ObjectBehavior;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\HttpKernel\KernelEvents;
@@ -17,7 +27,7 @@ class TranslatableLocaleListenerSpec extends ObjectBehavior
 
     function it_is_event_subscriber()
     {
-        $this->shouldBeAnInstanceOf('\Symfony\Component\EventDispatcher\EventSubscriberInterface');
+        $this->shouldBeAnInstanceOf(EventSubscriberInterface::class);
     }
 
     function it_handles_on_kernel_request_event()
